updates tests to lab/code
Removes devDependencies:

- jshint
- istanbul
- mocha
- chai

Adds devDependencies:

- lab@5.x.x
- code@1.x.x

Updated tests to work with lab and updated travis configuration to
push coverage to code climate. This includes the lab lint process
using the current Modulus .eslintrc until the official one is
published.
