+++ .eslintrc
@@ -1,0 +1,210 @@
+{
+  "ecmaFeatures": {
+    "arrowFunctions": false,
+    "binaryLiterals": true,
+    "blockBindings": true,
+    "classes": true,
+    "defaultParams": false,
+    "destructuring": false,
+    "forOf": true,
+    "generators": true,
+    "modules": false,
+    "objectLiteralComputedProperties": false,
+    "objectLiteralDuplicateProperties": false,
+    "objectLiteralShorthandMethods": true,
+    "objectLiteralShorthandProperties": true,
+    "octalLiterals": true,
+    "regexUFlag": false,
+    "regexYFlag": false,
+    "restParams": false,
+    "spread": false,
+    "superInFunctions": false,
+    "templateStrings": true,
+    "unicodeCodePointEscapes": false,
+    "globalReturn": false,
+    "jsx": false
+  },
+
+  "env": {
+    "browser": false,
+    "worker": false,
+    "node": true,
+    "amd": false,
+    "mocha": false,
+    "jasmine": false,
+    "phantomjs": false,
+    "jquery": false,
+    "prototypejs": false,
+    "shelljs": false,
+    "meteor": false,
+    "mongo": false,
+    "applescript": false,
+    "es6": false
+  },
+
+  "globals": {},
+
+  "plugins": [],
+
+  "rules": {
+    "comma-dangle": [2, "never"],
+    "no-cond-assign": 2,
+    "no-console": 1,
+    "no-constant-condition": 2,
+    "no-control-regex": 2,
+    "no-debugger": 2,
+    "no-dupe-args": 2,
+    "no-dupe-keys": 2,
+    "no-duplicate-case": 2,
+    "no-empty-character-class": 2,
+    "no-empty": 2,
+    "no-ex-assign": 2,
+    "no-extra-boolean-cast": 2,
+    "no-extra-parens": 2,
+    "no-extra-semi": 2,
+    "no-func-assign": 2,
+    "no-inner-declarations": [2, "functions"],
+    "no-invalid-regexp": 2,
+    "no-irregular-whitespace": 2,
+    "no-negated-in-lhs": 2,
+    "no-obj-calls": 2,
+    "no-regex-spaces": 2,
+    "no-reserved-keys": 0,
+    "no-sparse-arrays": 2,
+    "no-unreachable": 2,
+    "use-isnan": 2,
+    "valid-jsdoc": 0,
+    "valid-typeof": 2,
+    "no-unexpected-multiline": 2,
+    "accessor-pairs": 0,
+    "block-scoped-var": 0,
+    "complexity": 0,
+    "consistent-return": 0,
+    "curly": [2, "multi-line"],
+    "default-case": 0,
+    "dot-notation": 0,
+    "dot-location": 0,
+    "eqeqeq": 0,
+    "guard-for-in": 0,
+    "no-alert": 0,
+    "no-caller": 0,
+    "no-div-regex": 0,
+    "no-else-return": 0,
+    "no-empty-label": 0,
+    "no-eq-null": 0,
+    "no-eval": 0,
+    "no-extend-native": 0,
+    "no-extra-bind": 0,
+    "no-fallthrough": 0,
+    "no-floating-decimal": 0,
+    "no-implied-eval": 0,
+    "no-iterator": 0,
+    "no-labels": 0,
+    "no-lone-blocks": 0,
+    "no-loop-func": 0,
+    "no-multi-spaces": 0,
+    "no-multi-str": 0,
+    "no-native-reassign": 0,
+    "no-new-func": 0,
+    "no-new-wrappers": 0,
+    "no-new": 0,
+    "no-octal-escape": 0,
+    "no-octal": 0,
+    "no-param-reassign": 0,
+    "no-process-env": 0,
+    "no-proto": 0,
+    "no-redeclare": 0,
+    "no-return-assign": 0,
+    "no-script-url": 0,
+    "no-self-compare": 0,
+    "no-sequences": 0,
+    "no-throw-literal": 0,
+    "no-unused-expressions": 0,
+    "no-void": 0,
+    "no-warning-comments": 0,
+    "no-with": 0,
+    "radix": 0,
+    "vars-on-top": 0,
+    "wrap-iife": 0,
+    "yoda": 0,
+    "strict": 0,
+    "no-catch-shadow": 0,
+    "no-delete-var": 0,
+    "no-label-var": 0,
+    "no-shadow": 0,
+    "no-shadow-restricted-names": 0,
+    "no-undef": 0,
+    "no-undef-init": 0,
+    "no-undefined": 0,
+    "no-unused-vars": 0,
+    "no-use-before-define": 0,
+    "handle-callback-err": 0,
+    "no-mixed-requires": 0,
+    "no-new-require": 0,
+    "no-path-concat": 0,
+    "no-process-exit": 0,
+    "no-restricted-modules": 0,
+    "no-sync": 0,
+    "array-bracket-spacing": 0,
+    "brace-style": 0,
+    "camelcase": 0,
+    "comma-spacing": 0,
+    "comma-style": 0,
+    "computed-property-spacing": 0,
+    "consistent-this": 0,
+    "eol-last": 0,
+    "func-names": 0,
+    "func-style": 0,
+    "indent": 0,
+    "key-spacing": 0,
+    "lines-around-comment": 0,
+    "linebreak-style": 0,
+    "max-nested-callbacks": 0,
+    "new-cap": 0,
+    "new-parens": 0,
+    "new-parens": 0,
+    "newline-after-var": 0,
+    "no-array-constructor": 0,
+    "no-continue": 0,
+    "no-inline-comments": 0,
+    "no-lonely-if": 0,
+    "no-mixed-spaces-and-tabs": 0,
+    "no-multiple-empty-lines": 0,
+    "no-nested-ternary": 0,
+    "no-new-object": 0,
+    "no-spaced-func": 0,
+    "no-ternary": 0,
+    "no-trailing-spaces": 0,
+    "no-underscore-dangle": 2,
+    "one-var": 0,
+    "operator-assignment": 0,
+    "operator-linebreak": 0,
+    "padded-blocks": 0,
+    "quote-props": 0,
+    "quotes": 0,
+    "semi-spacing": 0,
+    "semi": 0,
+    "sort-vars": 0,
+    "space-after-keywords": 0,
+    "space-before-blocks": 0,
+    "space-before-function-paren": 0,
+    "space-in-parens": 0,
+    "space-infix-ops": 0,
+    "space-return-throw-case": 0,
+    "space-unary-ops": 0,
+    "spaced-comment": 0,
+    "wrap-regex": 0,
+    "constructor-super": 0,
+    "generator-star-spacing": 0,
+    "no-this-before-super": 0,
+    "no-var": 0,
+    "object-shorthand": 0,
+    "prefer-const": 1,
+    "max-depth": 0,
+    "max-len": 0,
+    "max-params": 0,
+    "max-statements": 0,
+    "no-bitwise": 0,
+    "no-plusplus": 0
+  }
+}
