--- .jshintrc
@@ -1,65 +1,0 @@
-{
-  "maxerr"        : 50,
-
-  "bitwise"       : true,
-  "camelcase"     : false,
-  "curly"         : false,
-  "eqeqeq"        : true,
-  "forin"         : true,
-  "immed"         : false,
-  "indent"        : 2,
-  "latedef"       : true,
-  "newcap"        : true,
-  "noarg"         : true,
-  "noempty"       : true,
-  "nonew"         : true,
-  "plusplus"      : false,
-  "quotmark"      : true,
-
-  "undef"         : true,
-  "unused"        : true,
-  "strict"        : false,
-  "trailing"      : true,
-
-  "asi"           : false,
-  "boss"          : false,
-  "debug"         : false,
-  "eqnull"        : true,
-  "es5"           : false,
-  "esnext"        : true,
-  "moz"           : false,
-
-  "evil"          : false,
-  "expr"          : true,
-  "funcscope"     : false,
-  "globalstrict"  : false,
-  "iterator"      : false,
-  "lastsemic"     : false,
-  "laxbreak"      : false,
-  "laxcomma"      : true,
-  "loopfunc"      : false,
-  "multistr"      : false,
-  "proto"         : false,
-  "scripturl"     : false,
-  "smarttabs"     : false,
-  "shadow"        : false,
-  "sub"           : false,
-  "supernew"      : false,
-  "validthis"     : false,
-
-  "node"          : true,
-
-  "globals"       : {
-    "afterEach"   : false,
-    "beforeEach"  : false,
-    "after"       : false,
-    "before"      : false,
-    "describe"    : false,
-    "it"          : false,
-    "spyOn"       : false,
-    "xit"         : false,
-    "xdescribe"   : false,
-    "window"      : false,
-    "document"    : false
-  }
-}
