+++ test/fixtures/npm.json
